Add files via upload
I realized last night that I created / was initializing the linked lists wrong but it was 3am at that point. I plan to do more but promised to upload to Github. I still have issue_requests, loading functionality, redoing initializations, and the proc_file to program before we start adding the mutexes and locks.